inference/requirements.txt
Fastapi==0.109.2
uvicorn==0.23.2
asyncio==3.4.3
python_dateutil==2.8.2
aiohttp==3.9.3
aiohttp_sse_client==0.2.1
jsonschema==4.20.0
PyJWT==2.8.0
pycryptodome==3.20.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-rerunfailures==13.0
PyYAML==6.0.1
pytest-assume==2.4.3
pytest-asyncio==0.21.1
pytest-ordering==0.6
allure-pytest==2.13.2
pytest-xdist==3.5.0
requests==2.31.0
aiobotocore==2.11.2
starlette-prometheus==0.9.0
tiktoken==0.5.2
numpy==1.26.4
